% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSeqpat.R
\name{CSeqpat}
\alias{CSeqpat}
\title{Mining Frequent Contiguous Sequential Patterns in a Text Corpus}
\usage{
CSeqpat(filepath, minsupport, docdelim)
}
\arguments{
\item{filepath}{Path to the text file/text corpus}

\item{minsupport}{Minimum absolute support for mining the patterns}

\item{docdelim}{Document delimiter in the corpus}
}
\value{
A dataframe containing the frequent phrase patterns with their absolute support
}
\description{
Takes in the filepath and minimum support and performs pattern mining
}
\examples{
test1 <- c("hoagie institution food year year road ",
"place little dated opened weekend fresh food")
tf <- tempfile()
writeLines(test1, tf)
CSeqpat(tf, 2, "\\t")
}
